pyproject.toml
[tool.poetry]
name = "stores-restful-api"
version = "0.1.0"
description = ""
authors = ["juuliadidenko <hiljaotsola2766@gmail.com>"]
readme = "README.md"
packages = [{include = "stores_restful_api"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.2"
flask-restful = "^0.3.9"
flask-jwt = "^0.3.2"
python-dotenv = "^0.21.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
